+++ src/compas/datastructures/mesh/subdivision.py
@@ -7,6 +7,10 @@ from math import pi
 from copy import deepcopy
 
 from compas.geometry import centroid_points
+from compas.geometry import offset_polygon
+
+from compas.utilities import iterable_like
+from compas.utilities import pairwise
 
 from compas.datastructures.mesh._mesh import Mesh
 from compas.datastructures.mesh.operations import mesh_split_edge
@@ -19,6 +23,7 @@ __all__ = [
     'mesh_subdivide_quad',
     'mesh_subdivide_catmullclark',
     'mesh_subdivide_doosabin',
+    'mesh_subdivide_frames',
     'trimesh_subdivide_loop',
 ]
 
@@ -76,7 +81,6 @@ class SubdMesh(Mesh):
             self.add_face([u, v, w])
         del self.face[fkey]
         return w
-
 
 # distinguish between subd of meshes with and without boundary
 # closed vs. open
@@ -203,6 +207,7 @@ def mesh_subdivide_corner(mesh, k=1):
     cls = type(mesh)
     for _ in range(k):
         subd = mesh_fast_copy(mesh)
+
         # split every edge
         for u, v in list(subd.edges()):
             mesh_split_edge(subd, u, v, allow_boundary=True)
@@ -472,6 +477,70 @@ def mesh_subdivide_doosabin(mesh, k=1, fixed=None):
     return subd2
 
 
+def mesh_subdivide_frames(mesh, offset, add_windows=False):
+    """Subdivide a mesh by creating offset frames and windows on its faces.
+
+    Parameters
+    ----------
+    mesh : Mesh
+        The mesh object to be subdivided.
+    offset : float or dict
+        The offset distance to create the frames.
+        A single value will result in a constant offset everywhere.
+        A dictionary mapping facekey: offset will be processed accordingly.
+    add_windows : boolean
+        Optional. Flag to add window face. Default is ``False``.
+
+    Returns
+    -------
+    Mesh
+        A new subdivided mesh.
+    """
+
+    subd = SubdMesh()
+
+    # 0. pre-compute offset distances
+    if not isinstance(offset, dict):
+        distances = iterable_like(mesh.faces(), [offset], offset)
+        offset = {fkey: od for fkey, od in zip(mesh.faces(), distances)}
+
+    # 1. add vertices
+    newkeys = {}
+    for vkey, attr in mesh.vertices(True):
+        newkeys[vkey] = subd.add_vertex(*mesh.vertex_coordinates(vkey))
+
+    # 2. add faces
+    for fkey in mesh.faces():
+        face = [newkeys[vkey] for vkey in mesh.face_vertices(fkey)]
+        d = offset.get(fkey)
+
+        # 2a. add face and break if no offset is found
+        if d is None:
+            subd.add_face(face)
+            continue
+
+        polygon = offset_polygon(mesh.face_coordinates(fkey), d)
+
+        # 2a. add offset vertices
+        window = []
+        for xyz in polygon:
+            x, y, z = xyz
+            new_vkey = subd.add_vertex(x=x, y=y, z=z)
+            window.append(new_vkey)
+
+        # 2b. frame faces
+        face = face + face[:1]
+        window = window + window[:1]
+        for sa, sb in zip(pairwise(face), pairwise(window)):
+            subd.add_face([sa[0], sa[1], sb[1], sb[0]])
+
+        # 2c. window face
+        if add_windows:
+            subd.add_face(window)
+
+    return subd
+
+
 def trimesh_subdivide_loop(mesh, k=1, fixed=None):
     """Subdivide a triangle mesh using the Loop algorithm.
 
@@ -609,7 +678,6 @@ def trimesh_subdivide_loop(mesh, k=1, fixed=None):
     subd2 = cls()
     subd2.data = subd.data
     return subd2
-
 
 # ==============================================================================
 # Main